Rename: OfficeViewModel -> PlaceViewModel
